refactor: rename address to bech32
