+++ packages/bech32/src/bech32.ts
