+++ packages/bech32/src/index.ts
@@ -1,0 +1,1 @@
+export * from './bech32';
