+++ packages/bech32/tsconfig.json
