+++ packages/contract/package.json
@@ -31,7 +31,7 @@
     "@ethersproject/logger": "^5.7.0",
     "@ethersproject/sha2": "^5.7.0",
     "@fuel-ts/abi-coder": "workspace:*",
-    "@fuel-ts/address": "workspace:*",
+    "@fuel-ts/bech32": "workspace:*",
     "@fuel-ts/math": "workspace:*",
     "@fuel-ts/interfaces": "workspace:*",
     "@fuel-ts/keystore": "workspace:*",
