+++ packages/fuels/src/index.ts
@@ -1,6 +1,6 @@
 export * from '@ethersproject/bytes';
 export * from '@fuel-ts/abi-coder';
-export * from '@fuel-ts/address';
+export * from '@fuel-ts/bech32';
 export * from '@fuel-ts/constants';
 export * from '@fuel-ts/contract';
 export * from '@fuel-ts/predicate';
