+++ packages/wallet/src/account.test.ts
@@ -1,5 +1,5 @@
 import { arrayify } from '@ethersproject/bytes';
-import { Bech32 } from '@fuel-ts/address';
+import { Bech32 } from '@fuel-ts/bech32';
 import { bn } from '@fuel-ts/math';
 import type {
   CallResult,
